OTELCollectorChanges
OTELCollectorChanges

+++ server/controller/controller.js
@@ -1,23 +1,26 @@
+const api = require('@opentelemetry/api');
+const tracer = require('../../tracing')('anand-example');
+const http = require('http');
+
 var Userdb = require('../model/model');
 
 // create and save new user
 exports.create = (req,res)=>{
     // validate request
 
-    console.log("Inside Create Span\n\n")
-    const api = require('@opentelemetry/api');
-    let current_span = api.trace.getSpan(api.context.active());
-    let trace_id = current_span.spanContext().traceId;
-    let span_id = current_span.spanContext().spanId;
-    let trace_flags = current_span.spanContext().traceFlags;
-    let is_remote = current_span.spanContext().isRemote;
-
-    console.log(current_span)
-    console.log(`log trace_id:”${is_remote}”`)
-    console.log(`log trace_id:”${trace_id}” span_id:”${span_id}” trace_flags:”${trace_flags}”`);
-
+    console.log("\n\nInside create Span \n\n")
+    console.log("Inside req header:"+ JSON.stringify(req.headers))
+    const currentSpan = api.trace.getActiveSpan();
+    // display traceid in the terminal
+    const traceId = currentSpan.spanContext().traceId;
+    console.log(`traceId: ${traceId}`);
+    const span = tracer.startSpan('handleRequest', {
+      kind: 1, // server
+      attributes: { key: 'value' },
+    });
+    // Annotate our span to capture metadata about the operation
+    span.addEvent('invoking create request');
     
-    console.log("End of Create Span\n\n")
 
     if(!req.body){
         res.status(400).send({ message : "Content can not be emtpy!"});
@@ -44,26 +47,26 @@ exports.create = (req,res)=>{
                 message : err.message || "Some error occurred while creating a create operation"
             });
         });
-
+    span.end();
 }
 
 // retrieve and return all users/ retrive and return a single user
 exports.find = (req, res)=>{
 
-    console.log("\n\nInside Find Span \n\n")
-    const api = require('@opentelemetry/api');
-    let current_span = api.trace.getSpan(api.context.active());
-    let trace_id = current_span.spanContext().traceId;
-    let span_id = current_span.spanContext().spanId;
-    let trace_flags = current_span.spanContext().traceFlags;
-    let is_remote = current_span.spanContext().isRemote;
 
-    console.log(current_span)
-    console.log(`log trace_id:”${is_remote}”`)
-    console.log(`log trace_id:”${trace_id}” span_id:”${span_id}” trace_flags:”${trace_flags}”`);
-
+    console.log("\n\nInside find Span \n\n")
+    console.log("Inside req header:"+ JSON.stringify(req.headers))
+    const currentSpan = api.trace.getActiveSpan();
+    // display trace id in the terminal
+    const traceId = currentSpan.spanContext().traceId;
+    console.log(`traceId: ${traceId}`);
+    const span = tracer.startSpan('handleRequest', {
+      kind: 1, // server
+      attributes: { key: 'value' },
+    });
+    // Annotate our span to capture metadata about the operation
+    span.addEvent('invoking find request');
     
-    console.log("\n\nEnd of Find Span \n\n")
 
     if(req.query.id){
         const id = req.query.id;
@@ -90,26 +93,24 @@ exports.find = (req, res)=>{
             })
     }
 
-    
+    span.end();
 }
 
 // Update a new idetified user by user id
 exports.update = (req, res)=>{
 
-    console.log("\n\nInside update Span \n\n")
-    const api = require('@opentelemetry/api');
-    let current_span = api.trace.getSpan(api.context.active());
-    let trace_id = current_span.spanContext().traceId;
-    let span_id = current_span.spanContext().spanId;
-    let trace_flags = current_span.spanContext().traceFlags;
-    let is_remote = current_span.spanContext().isRemote;
-
-    console.log(current_span)
-    console.log(`log trace_id:”${is_remote}”`)
-    console.log(`log trace_id:”${trace_id}” span_id:”${span_id}” trace_flags:”${trace_flags}”`);
-
-    
-    console.log("\n\nEnd of update Span \n\n")
+    console.log("\n\nInside Update Span \n\n")
+    console.log("Inside req header:"+ JSON.stringify(req.headers))
+    const currentSpan = api.trace.getActiveSpan();
+    // display traceid in the terminal
+    const traceId = currentSpan.spanContext().traceId;
+    console.log(`traceId: ${traceId}`);
+    const span = tracer.startSpan('handleRequest', {
+      kind: 1, // server
+      attributes: { key: 'value' },
+    });
+    // Annotate our span to capture metadata about the operation
+    span.addEvent('invoking update request');
     
     if(!req.body){
         return res
@@ -130,25 +131,35 @@ exports.update = (req, res)=>{
         .catch(err =>{
             res.status(500).send({ message : "Error Update user information"})
         })
+
+    span.end();
 }
 
 // Delete a user with specified user id in the request
 exports.delete = (req, res)=>{
 
     console.log("\n\nInside Delete Span \n\n")
-    const api = require('@opentelemetry/api');
-    let current_span = api.trace.getSpan(api.context.active());
-    let trace_id = current_span.spanContext().traceId;
-    let span_id = current_span.spanContext().spanId;
-    let trace_flags = current_span.spanContext().traceFlags;
-    let is_remote = current_span.spanContext().isRemote;
-
-    console.log(current_span)
-    console.log(`log trace_id:”${is_remote}”`)
-    console.log(`log trace_id:”${trace_id}” span_id:”${span_id}” trace_flags:”${trace_flags}”`);
-
+    console.log("Inside req header:"+ JSON.stringify(req.headers))
     
-    console.log("\n\nEnd of Delete Span \n\n")
+
+
+    const currentSpan = api.trace.getActiveSpan();
+    // display traceid in the terminal
+    const traceId = currentSpan.spanContext().traceId;
+    console.log(`traceId: ${traceId}`);
+    const span = tracer.startSpan('handleRequest', {
+      kind: 1, // server
+      attributes: { key: 'value' },
+    });
+    // Annotate our span to capture metadata about the operation
+    span.addEvent('invoking delete request');
+    
+
+    if(!req.body){
+        return res
+            .status(400)
+            .send({ message : "Data to update can not be empty"})
+    }
 
     const id = req.params.id;
     Userdb.findByIdAndDelete(id)
@@ -166,4 +177,5 @@ exports.delete = (req, res)=>{
                 message: "Could not delete User with id=" + id
             });
         });
+    span.end();
 }
